// src/api/participants/constants.ts
export enum ACTIONS_TYPES {
    POST_PARTICIPANT_TO_EVENT = 'POST_PARTICIPANT_TO_EVENT',
    POST_PARTICIPANT_TO_EVENT_SUCCEED = 'POST_PARTICIPANT_TO_EVENT_SUCCEED',

    PUT_PARTICIPANT_TO_EVENT = 'PUT_PARTICIPANT_TO_EVENT',
    PUT_PARTICIPANT_TO_EVENT_SUCCEED = 'PUT_PARTICIPANT_TO_EVENT_SUCCEED',

    REMOVE_PARTICIPANT_FROM_EVENT = 'REMOVE_PARTICIPANT_FROM_EVENT',
    REMOVE_PARTICIPANT_FROM_EVENT_SUCCEED = 'REMOVE_PARTICIPANT_FROM_EVENT_SUCCEED',

    POST_PARTICIPANT_TO_TRAVEL = 'POST_PARTICIPANT_TO_TRAVEL',
    POST_PARTICIPANT_TO_TRAVEL_SUCCEED = 'POST_PARTICIPANT_TO_TRAVEL_SUCCEED',

    PUT_PARTICIPANT_TO_TRAVEL = 'PUT_PARTICIPANT_TO_TRAVEL',
    PUT_PARTICIPANT_TO_TRAVEL_SUCCEED = 'PUT_PARTICIPANT_TO_TRAVEL_SUCCEED',

    REMOVE_PARTICIPANT_FROM_TRAVEL = 'REMOVE_PARTICIPANT_FROM_TRAVEL',
    REMOVE_PARTICIPANT_FROM_TRAVEL_SUCCEED = 'REMOVE_PARTICIPANT_FROM_TRAVEL_SUCCEED',
}